Refactor parsing functions into submodule
